requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
LEPL==5.1.3
MarkupSafe==2.0.1
pyrus-api==1.23.0
requests==2.25.1
rfc6266==0.0.4
urllib3==1.26.5
Werkzeug==2.0.1
